+++ CHANGELOG.md
@@ -21,6 +21,7 @@ the code was deployed.
 - Updated edit client dashboard page to add new status and first_device_live_at fields (CU-86dv9uych).
 - Updated submitUpdateClientTest.js test cases with new client fields (CU-86dv9uych).
 - Updated dashboard homepage to include and sort by Organization (CU-86dva4dxd).
+- Fixed error caused due to updating client's first_device_live_at in dashboard (CU-)
 
 ## [14.1.0] - 2024-11-14